fix folder deletion bug (#715)
prevent deletion of folders that are NOT actually subfolders and only
start with the same sequence of characters. Originally, if I have two
or more folders on the same level:
`slider`
`slider-homepage`
and I delete only the `slider` one, `slider-homepage` would get deleted
as well.

This change makes sure the deleted folders are actually subfolders of
the deleted one.